Add files via upload
In the correlation part, webcam gaze tracking along the horizontal dimension was strongest in Sections 2, 5, and 8 (the vertical centre of the screen), and surprisingly the y correlation of Sections 1 and 2 is negative.

In the absolute distance part, the results were stable, with the largest difference in section 123 (top screen) and the smallest difference in section 789 (bottom screen). However, this difference may be due to the visual angle.

Can discuss the options for visualization.